django/todo_proj/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
psycopg==3.1.18
psycopg-binary==3.1.18
Pygments==2.3.1
python-dotenv==1.0.1
pytz==2018.9
redgreenunittest==0.1.1
sqlparse==0.4.4
typing_extensions==4.10.0